Implement Alert component with ReactDOM createPortal
