+++ src/components/Base/Util/Variants/alert.ts
@@ -1,0 +1,13 @@
+export const alertVariant = {
+  init: {
+    opacity: 0,
+    scale: 0.5,
+  },
+  open: {
+    opacity: 1,
+  },
+  exit: {
+    opacity: 0,
+    scale: 0.5,
+  },
+};
